refactor(backend): remove strict libcc
